Add files via upload
Scripts v0 - v6 reflect the editing history of the main script. Coefficients 4 - 6 are the lists of predictors and their associated coefficients with scripts v4 - v6 respectively. The Bond Length Predictor allows for a single prediction to be made about the bond length given the refcode and the predictor values. Heatmap creates a quick visualization of the strength of a predictor on the final bond length prediction.